refactor add_tones_to_word and expand tests in their own file
